Move Logos into a branding app
