+++ website/website/templates/base.html
@@ -1,4 +1,4 @@
-{% load i18n static compress wagtailcore_tags wagtailuserbar wagtailsettings_tags gravatar site_tags %}
+{% load i18n static compress wagtailcore_tags wagtailuserbar wagtailsettings_tags gravatar site_tags organisation_menu_tags %}
 {% get_current_language as LANGUAGE_CODE %}
 {% get_site_root as site_root %}
 {% get_settings %}
