+++ website/website/templatetags/site_tags.py
@@ -1,7 +1,5 @@
 from django import template
 
-from website.models import Logo
-
 register = template.Library()
 
 
@@ -50,16 +48,4 @@ def main_menu_children(context, parent):
         'menuitems_children': menuitems_children,
         # required by the pageurl tag that we want to use within this template
         'request': context['request'],
-    }
-
-
-@register.inclusion_tag('tags/organisation_menu.html', takes_context=True)
-def organisation_menu(context, color=''):
-    logos = Logo.objects.exclude(belongs_to=context['request'].site).all()
-    committees = logos.filter(category='committee')
-    sections = logos.filter(category='section')
-    return {
-        'committees': committees,
-        'sections': sections,
-        'color': color,
     }
